requirements.txt
streamlit
langchain
langchain-groq
langchain-community
chromadb
huggingface-hub
sentence-transformers
python-dotenv
bs4
requests
newspaper3k
